+++ Sofa/Component/Mapping/NonLinear/CMakeLists.txt
@@ -8,8 +8,8 @@ set(HEADER_FILES
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/init.h
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/AreaMapping.h
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/AreaMapping.inl
-    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/AssembledNonLinearMapping.h
-    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/AssembledNonLinearMapping.inl
+    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/BaseNonLinearMapping.h
+    ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/BaseNonLinearMapping.inl
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceFromTargetMapping.h
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceFromTargetMapping.inl
     ${SOFACOMPONENTMAPPINGNONLINEAR_SOURCE_DIR}/DistanceMapping.h
